Componentize
- Add Blue Sky rough poc to component.
