+++ pages/index.vue
@@ -1,20 +1,4 @@
 <script lang="ts" setup>
-import { ref, onMounted } from 'vue';
-
-const posts = ref<any[]>([]);
-const loading = ref(true);
-const error = ref<string | null>(null);
-
-onMounted(async () => {
-  try {
-    posts.value = await $fetch('/api/fetch-bsky');
-  } catch (err: any) {
-    error.value = err.message;
-  } finally {
-    loading.value = false;
-  }
-});
-
 const { data: blogs } = await useAsyncData('blog', () =>
   queryContent('/blog').find()
 );
@@ -22,14 +6,6 @@ const { data: blogs } = await useAsyncData('blog', () =>
 useHead({
   title: 'The Blogroject',
 });
-
-const convertDate = (date: string) => {
-  return new Date(date).toLocaleDateString('en-US', {
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric',
-  });
-};
 </script>
 
 <template>
@@ -40,29 +16,9 @@ const convertDate = (date: string) => {
           <h2>
             <NuxtLink :to="blog._path">{{ blog.title }}</NuxtLink>
           </h2>
-
           <p>{{ blog.description }}</p>
         </li>
-        <!-- Componentize eventually -->
-        <li class="bsky-post" v-if="loading">Loading...</li>
-        <li class="bsky-post" v-if="error">{{ error }}</li>
-        <li class="bsky-post" v-if="!loading && !error && !posts.length">
-          No posts found.
-        </li>
-        <li
-          class="bsky-post"
-          v-if="!loading && !error && posts.length"
-          v-for="post in posts"
-        >
-          <img
-            style="display: none"
-            :src="post.post.author.avatar"
-            :alt="post.post.author.displayName"
-          />
-          <h2>@{{ post.post.author.handle }}</h2>
-          <p>{{ post.post.record.text }}</p>
-          <p>{{ convertDate(post.post.record.createdAt) }}</p>
-        </li>
+        <BlueSky />
       </ul>
     </section>
   </main>
